requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
cytoolz==0.12.0
defusedxml==0.7.1
Django==4.1.2
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
inflection==0.5.1
ipfshttpclient==0.8.0a2
jsonschema==4.16.0
lru-dict==1.1.8
Markdown==3.4.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
oauthlib==3.2.2
parsimonious==0.8.1
protobuf==3.19.5
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.6.0
pyotp==2.7.0
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.5
PyYAML==6.0
razorpay==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rlp==2.0.1
six==1.16.0
sqlparse==0.4.3
toolz==0.12.0
uritemplate==4.1.1
urllib3==1.26.12
varint==1.0.2
web3==5.31.1
websockets==9.1
whitenoise==6.2.0
yarl==1.8.1
